chore: REPORTED 추가하기
- 카드 상태 이상하게 나오던것 해결
